yolov7_small_objects_modules/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.0.4
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.0
idna==3.10
imagecodecs==2024.12.30
imageio==2.37.0
imantics==0.1.12
ipython==8.29.0
jedi==0.19.2
jinja2==3.1.4
kiwisolver==1.4.7
lazy-loader==0.4
lxml==5.3.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.2
numpy==1.23.5
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt-toolkit==3.0.48
protobuf==4.21.2
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pybboxes==0.1.6
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywavelets==1.8.0
PyYAML==6.0.2
requests==2.32.3
sahi==0.11.20
scikit-image==0.19.3
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.7
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2025.2.18
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.6
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.25
ultralytics-thop==2.0.10
urllib3==2.2.3
wcwidth==0.2.13
werkzeug==3.1.3
xmljson==0.2.1
